src/requirements.txt

python-telegram-bot==13.0
pymongo==3.11.1
python-dotenv==0.15.0
# lambdarest==10.1.0
# Werkzeug
# pyrsistent
# .